Split up `RDDSpec`
